Separate schema and resolver of accounts from karkun
