+++ idreesia-common/server/schemas/admin/index.js
@@ -1,3 +1,1 @@
-export { default as AdminJob } from "./admin-job";
-export { default as Profile } from "./profile";
-export { default as Department } from "./department";
+export { default as AdminJob } from './admin-job';
