+++ idreesia-web/imports/ui/modules/admin/security-accounts/index.js
@@ -1,3 +1,3 @@
 export { default as AccountsList } from './list';
 export { default as AccountsNewForm } from './new-form';
-export { default as AccountsEditForm } from './edit-form';
+export { default as AccountsEditForm } from './edit/edit-form';
